app/requirements.txt
# devops
black
click
pylint
pytest
pytest-cov

# web app
python-dotenv
requests
streamlit
spotipy
pydub
pyaudio
openai
wave
whisper